// src/types/api/payment.d.ts
export namespace Payment {
    /**
     * Payment Object
     */
    export type PaymentObject = {
        /**
         * Shop ID that registered this payment.
         */
        shop_id: string;

        /**
         * Payment ID. (sometimes called "Order ID")
         */
        id: string;

        /**
         * Accepted payment method in this payment. 
         * 
         * - `Card`: this payment accepts card.
         */
        pay_type: "Card";
// ...
        access_id: string;
// ...
        customer_id: string | null;
// ...
        card_id: string | null;
// ...
    }
// ...
}